Make the id manager refuse rather than guess, and close the last sweeps
IdManager sat in the flagged-and-unchanged list on the argument that
adding a way to abort startup is the server owner's decision. That
argument still holds, and no abort was needed.

The damaging failure was the second of two. Each of the five queries that
collect used object ids runs on its own thread with a catch that logged
and swallowed, so a query that did not finish silently removed a whole
table's ids from the set, and getUsedIds returned a partial result
indistinguishable from a complete one. The server then handed out ids
that live rows already held: object id collisions, quietly, of the kind
that cannot be undone. getUsedIds now records whether any query failed,
did not finish, or was interrupted, and throws instead of returning a
partial set, because a partial set is worse than none -- it looks whole.
IdManager sets _initialized only at the end of a complete build, and
getNextId throws while it is false. An immediate failure with a message,
instead of quiet corruption.

Two sweeps close the list. Only two classes define compareTo and neither
defines equals, but two cases are not a population to diverge from, and
consistency with equals only bites in a sorted collection that uses
compareTo in place of equals; neither appears in a TreeSet or TreeMap.

Of 69 divisions whose result lands in a double or float, the first pass
flagged 24 and 23 were my own false positives: the heuristic does not
resolve types, so it flagged expressions whose operands are already
double. Checked one by one, distance comes from Math.hypot, getChance
returns double, sqLevel is declared double, and milliToStart / 1000 is a
deliberate integer division into seconds. One survived. In RecipeManager
the per-pass HP or MP cost of a recipe divided an int by an int and
assigned the truncated result to a double, so a cost of 3 spread over 5
passes charged nothing and any cost that did not divide evenly charged
short.

Core and datapack compile, 54 tests pass.

+++ java/org/l2jmobius/gameserver/managers/RecipeManager.java
@@ -531,7 +531,10 @@ public class RecipeManager
 			boolean ret = true;
 			for (RecipeStatHolder statUse : _recipeList.getStatUse())
 			{
-				final double modifiedValue = statUse.getValue() / _creationPasses;
+				// Both operands are ints, so this divided as an integer and the double only
+				// received the truncated result: a cost of 3 spread over 5 passes charged
+				// nothing at all, and every cost that did not divide evenly charged short.
+				final double modifiedValue = (double) statUse.getValue() / _creationPasses;
 				if (statUse.getType() == StatType.HP)
 				{
 					// we do not want to kill the player, so its CurrentHP must be greater than the reduce value
